Refactor Monte Carlo validation step with improved module structure
