+++ src/training/step_config.py
@@ -279,7 +279,7 @@ PIPELINE_STEPS: Dict[str, StepConfig] = {
         step_number="19",
         step_name="monte_carlo_validation",
         description="Validate models using Monte Carlo simulation",
-        module_path="src.training.steps.validation_components.monte_carlo_validation_step",
+        module_path="src.training.steps.model_training.validation.step19_monte_carlo_validation",
         class_name="MonteCarloValidationStep",
         dependencies=["18"],
         required_inputs=["optimized_models"],
